[DSL] Finally fixed typing problems in generated code, implemented all remaining M extension instructions, added tests for them. Everything works, however, there is one problem: signed division and remainder overflow which requires control flow in DSL to handle, which I have not implemented yet.
TODO:
0. Make pretty dump for CPU state.
1. JIT.
2. Decoder tree with correct algo
3. DSL syntax improvements
4. Improve README
5. Support more syscalls and extensions.
+ polish code

+++ Sim/include/ExecContext.hpp
@@ -31,8 +31,8 @@ public:
   virtual void syscall() = 0;
 
   virtual int bitrev(int Val, int NBits = 32) = 0;
-  virtual constexpr uint64_t sext(uint64_t Val, int N) = 0;
-  virtual constexpr uint64_t zext(uint64_t V, unsigned N) = 0;
+  virtual constexpr uint64_t sext(uint64_t Val, int64_t N) = 0;
+  virtual constexpr uint64_t zext(uint64_t V, int64_t N) = 0;
   virtual constexpr int saturateUnsigned(unsigned Val, unsigned N) = 0;
 };
 } // namespace GeneralSim
